+++ example/lib/main.dart
@@ -4,6 +4,7 @@ import 'package:adbrixrm_flutter/adbrixrm.dart';
 import 'package:adbrixrm_flutter_example/commerceEvent.dart';
 import 'package:adbrixrm_flutter_example/gameEvent.dart';
 import 'package:adbrixrm_flutter_example/userInfoEvent.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
@@ -17,9 +18,9 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
-
   String _deeferredDeeplink;
   String _deeplink;
+  bool _idfaAuthorize = false;
 
   int _currentIndex = 0;
   List _page = [userInfoView(), commerceView(), gameView()];
@@ -29,38 +30,44 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
     super.initState();
     WidgetsBinding.instance.addObserver(this);
 
-    AdBrixRm.sdkInit(appKey: 'G2Iz74fLkUOcZPZTrZQnQw', secretKey: 'ZP1RO2EDY02kpifcIOlzGQ');
+    TargetPlatform defaultTargetPlatform;
+
+    // set IDFA
+    if (defaultTargetPlatform == TargetPlatform.iOS) {
+      if (_idfaAuthorize = false) {
+        AdBrixRm.stopGettingIDFA();
+      } else {
+        AdBrixRm.startGettingIDFA();
+      }
+    }
+
+    AdBrixRm.sdkInit(
+        appKey: 'G2Iz74fLkUOcZPZTrZQnQw', secretKey: 'ZP1RO2EDY02kpifcIOlzGQ');
     AdBrixRm.setEventUploadCountInterval(
         interval: AdBrixEventUploadCountInterval.MIN);
     AdBrixRm.setEventUploadTimeInterval(
         interval: AdBrixEventUploadTimeInterval.MIN);
     AdBrixRm.setLogLevel(logLevel: AdBrixLogLevel.ERROR);
 
-    Timer(Duration(seconds: 3),() {
+    Timer(Duration(seconds: 3), () {
       getDeferredDeeplink();
       getDeeplink();
     });
-
   }
+
   @override
   void didChangeAppLifecycleState(AppLifecycleState currentState) {
     if (currentState == AppLifecycleState.resumed) {
-
       print("App is onResume");
       getDeeplink();
-
     }
-
   }
 
-  Future <void> getDeferredDeeplink () async {
-
+  Future<void> getDeferredDeeplink() async {
     String deferredDeeplink;
 
-    try{
-
+    try {
       deferredDeeplink = await AdBrixRm.adbrixDeferredDeeplink;
-
     } on PlatformException {
       print("there is no deferred deeplink");
     }
@@ -73,14 +80,12 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
       print("mydeferredDeeplink");
       print(_deeferredDeeplink);
     }
-
   }
 
-  Future <void> getDeeplink () async {
-
+  Future<void> getDeeplink() async {
     String deeplink;
 
-    try{
+    try {
       deeplink = await AdBrixRm.adbrixDeeplink;
     } on PlatformException {
       print("there is no deferred deeplink");
@@ -93,9 +98,7 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
       print("myDeeplink");
       print(_deeplink);
     }
-
   }
-
 
   @override
   Widget build(BuildContext context) {
@@ -112,12 +115,10 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
           currentIndex: _currentIndex,
           onTap: onTabTapped,
           items: const <BottomNavigationBarItem>[
+            BottomNavigationBarItem(icon: Icon(Icons.home), label: 'userinfo'),
             BottomNavigationBarItem(
-                icon: Icon(Icons.home), label: 'userinfo'),
-            BottomNavigationBarItem(
-                icon: Icon(Icons.credit_card), label:'Commerce'),
-            BottomNavigationBarItem(
-                icon: Icon(Icons.gamepad), label:'Game')
+                icon: Icon(Icons.credit_card), label: 'Commerce'),
+            BottomNavigationBarItem(icon: Icon(Icons.gamepad), label: 'Game')
           ],
         ),
       ),
@@ -134,7 +135,7 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
         Map<String, dynamic> attr = <String, dynamic>{
           'int': 2,
           'string': 'this is commerce',
-          'string2' : 'this is String2',
+          'string2': 'this is String2',
           'double': 25000.00,
           'long': 2222222222222,
           'bool': false
@@ -151,7 +152,6 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
           'double': 35000.00
         };
 
-
         AdBrixRm.events(eventName: 'game', attr: attr);
 
         break;
@@ -161,4 +161,3 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
     });
   }
 }
-